requirements.txt
PyQt5>=5.15.0
Pillow>=9.0.0
ffmpeg-python>=0.2.0
pyinstaller>=5.0.0 